update:
Instead of a modal window, I made a new page with a room

Then I finished it more or less, but it was still quite dryly done. The minimum you can do is throw an invite - it can be accepted and the game can start. Division into Creator and occupant :)