update invite user (Part 2)
Invited users to 90% functionality (Part 2)

Completely redesigned functionality. I removed the modal window and made a separate tab instead to make it easier to work. And I added functionality for sending a leader (for now only by leaving the room), invitations, returning to the room from home and leaving the room from there.