New draw conditions
*If a participant, having at the end of the game three kings against the opponent’s one king and at the same time owning the “high road”,
on his 15th move (counting from the moment the balance of forces is established) will not capture the opponent’s king

*If a participant has three kings or more checkers (simple checkers and kings) against one king of the opponent at the end of the game,
on his 15th move (counting from the moment the balance of forces is established) he will not capture the enemy’s king.

* 15 move rule

*If the same position (the same arrangement of checkers) is repeated three (or more) times,
Moreover, the turn to move will be on the same side each time.