update:
I adapted the board, added a button for the computer to return to the main menu and fixed a bug: if the user was looking at the bot's move and at that moment the user was looking at the history of the moves, it could be bugged. Now, if the user is watching the history of moves, the bot is waiting for the user to return to the current state.